model: split fips config helper
src/v/config depends on quite a few things in
src/v/model/{fundamental,metadata} unfortunately. so having fips_enable
from_config helper in model/fundamental.h creates a circular dependency.
the fix is to split out this one function into a separate header.
